:bug: fix(components): fix font colors and list styles in accordion list
