+++ libs/domains/blocks/feature/src/accordion-list/index.tsx
@@ -1,13 +1,13 @@
 import { PortableTextBlock } from 'next-sanity'
 
 import { AccordionListBlockType } from '@alecia/block-types'
-import BlogPortableText from '@alecia/blog/blog-portable-text'
 import {
   AccordionContent,
   AccordionItem,
   AccordionRoot as Accordion,
   AccordionTrigger,
 } from '@alecia/ui-kit/ui/accordion'
+import BasicPortableText from '@alecia/ui-kit/ui/portable-text/basic-portable-text'
 import { Typography } from '@alecia/ui-kit/ui/typography'
 
 // TODO: Add different layouts for the accordion list (centered, next to text, etc)
@@ -27,7 +27,7 @@ const AccordionList = ({ pretitle, heading, intro, items }: AccordionListBlockTy
             variant="p"
             className="text-primary-950 dark:text-primary-300 max-w-[300px] max-md:mx-auto lg:text-[1.125rem] lg:leading-relaxed"
           >
-            <BlogPortableText value={intro as PortableTextBlock[]} />
+            <BasicPortableText value={intro as PortableTextBlock[]} changeOnDarkMode />
           </Typography>
         ) : null}
       </div>
@@ -36,7 +36,7 @@ const AccordionList = ({ pretitle, heading, intro, items }: AccordionListBlockTy
           <AccordionItem key={item._key} value={item._key}>
             <AccordionTrigger className="text-base">{item.trigger}</AccordionTrigger>
             <AccordionContent className="">
-              <BlogPortableText value={item.content as PortableTextBlock[]} />
+              <BasicPortableText value={item.content as PortableTextBlock[]} />
             </AccordionContent>
           </AccordionItem>
         ))}
